fix: corrige price como decimal

+++ back-end/src/database/models/sales.js
@@ -25,7 +25,7 @@ module.exports = (sequelize, DataTypes) => {
         },
       },
       totalPrice: {
-        type: DataTypes.INTEGER,
+        type: DataTypes.DECIMAL(4, 2),
         allowNull: false,
       },
       deliveryAddress: {
